requirements.txt
Flask==2.2.0
pytesseract==0.3.8
Pillow==10.0.0
Werkzeug==2.2.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
